requirements.txt
asgiref==3.8.1
build==1.2.2.post1
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.9
Django==5.1.2
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
dnspython==2.7.0
filelock==3.16.1
identify==2.6.1
idna==3.10
nodeenv==1.9.1
packaging==24.1
pip-tools==7.4.1
pipenv==2024.1.0
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2-binary==2.9.10
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.2.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.6
wheel==0.44.0
